refactor(datatype): 将品牌单价数据类型从 float 改为 double
- 在 Brand 结构体中，将 price 字段的类型从 float 修改为 double
-这个改动可以提高价格数据的精度，避免因浮点数精度问题导致的计算误差

+++ include/datatype.h
@@ -19,7 +19,7 @@ struct Brand {
     std::string name; ///< 品牌名称（唯一标识）
     int code; ///< 品牌编码
     int quantity; ///< 当前库存数量
-    float price; ///< 单品价格
+    double price; ///< 单品价格
 
     bool operator==(const Brand& other) const {
         return code == other.code && name == other.name && quantity == other.quantity && price == other.price;
